+++ monai/apps/mmars/__init__.py
@@ -10,7 +10,7 @@
 # limitations under the License.
 
 from .config_parser import ConfigParser
-from .config_resolver import ConfigComponent, ConfigResolver, ModuleScanner
+from .config_resolver import ConfigComponent, ConfigResolver
 from .mmars import download_mmar, get_model_spec, load_from_mmar
 from .model_desc import MODEL_DESC, RemoteMMARKeys
-from .utils import get_class, instantiate_class, search_configs_with_objs, update_configs_with_objs
+from .utils import search_configs_with_deps, update_configs_with_deps
